feat(GIFT-13526): delete ods business centres endpoint
