+++ test/dom/dom.business-centres.api-test.ts
@@ -1,0 +1,42 @@
+import { HttpStatus } from '@nestjs/common';
+import AppConfig from '@ukef/config/app.config';
+import { Api } from '@ukef-test/support/api';
+
+const {
+  domOdsVersioning: { prefixAndVersion },
+} = AppConfig();
+
+describe('/dom - business centres', () => {
+  let api: Api;
+
+  beforeAll(async () => {
+    api = await Api.create();
+  });
+
+  afterAll(async () => {
+    await api.destroy();
+  });
+
+  describe('/business-centres', () => {
+    const url = `/api/${prefixAndVersion}/dom/business-centres`;
+
+    it(`should return ${HttpStatus.OK} with mapped business centres`, async () => {
+      // Act
+      const { status, body } = await api.get(url);
+
+      // Assert
+      expect(status).toBe(HttpStatus.OK);
+
+      const expected = expect.arrayContaining([
+        expect.objectContaining({
+          code: expect.any(String),
+          name: expect.any(String),
+        }),
+      ]);
+
+      expect(body).toEqual(expected);
+    });
+
+    // TODO: APIM-613 - create a mock request to mimic receiving a 500 error from ODS.
+  });
+});
