+++ test/ods/ods.business-centres.api-test.ts
@@ -18,29 +18,6 @@ describe('/ods - business centres', () => {
     await api.destroy();
   });
 
-  describe('/business-centres', () => {
-    const url = `/api/${prefixAndVersion}/ods/business-centres`;
-
-    it(`should return ${HttpStatus.OK} with mapped business centres`, async () => {
-      // Act
-      const { status, body } = await api.get(url);
-
-      // Assert
-      expect(status).toBe(HttpStatus.OK);
-
-      const expected = expect.arrayContaining([
-        expect.objectContaining({
-          code: expect.any(String),
-          name: expect.any(String),
-        }),
-      ]);
-
-      expect(body).toEqual(expected);
-    });
-
-    // TODO: APIM-613 - create a mock request to mimic receiving a 500 error from ODS.
-  });
-
   describe('/business-center/:code/non-working-days', () => {
     const baseUrl = `/api/${prefixAndVersion}/ods/business-centre`;
